feature: add 3rd scenario, new exception

+++ src/main/java/org/example/summerpracticesecondyear/repositories/MovieRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MovieRepository extends BaseRepository<Movie, Integer> {
+public interface MovieRepository extends BaseRepository<Movie, Long> {
 
     @Query(value = "select m.id as id, m.title as title, m.duration as duration, m.genre as genre from Movie m " +
             "where m.genre = :genre")
