+++ src/main/java/org/example/summerpracticesecondyear/repositories/RoomRepository.java
@@ -4,5 +4,5 @@ import org.example.summerpracticesecondyear.entities.Room;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomRepository extends BaseRepository<Room, Integer> {
+public interface RoomRepository extends BaseRepository<Room, Long> {
 }
